backend/requirements.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==11.3.0

# Auth
djangorestframework-simplejwt==5.3.1

# Schema / Docs
drf-spectacular==0.27.2

# DB
psycopg2-binary==2.9.11
django-extensions==3.2.3

# Storage (DigitalOcean Spaces / S3)
django-storages==1.14.4
boto3==1.35.0

# Async / WebSockets
channels==4.1.0
channels-redis==4.2.0
uvicorn[standard]==0.34.0

# Task Queue
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.1.1

# News
newsapi-python==0.2.7

# Search
django-filter==24.3

# Cache
django-redis==5.4.0

# Rate limiting
django-ratelimit==4.1.0

# Utilities
requests==2.32.3
setuptools>=69.0.0
